Created a better player prefab
The camera is a child of the Player GameObject once again, just to make it easier for the prefab to have all the components in one place.

The CameraControl script was rewritten a little bit so that it can be a component of the Player that also controls the camera's transform.